requirements.txt
python-dotenv
neon-api
psycopg2-binary
langchain
langgraph
langchain-google-genai